requirements.txt
matplotlib==3.1.2
numpy==1.16.4
Pillow==6.2.1
